+++ sdk/cosmos/azure-cosmos/azure/cosmos/_endpoint_discovery_retry_policy.py
@@ -73,9 +73,13 @@ class EndpointDiscoveryRetryPolicy(object):
         if self.request.location_endpoint_to_route:
             if _OperationType.IsReadOnlyOperation(self.request.operation_type):
                 # Mark current read endpoint as unavailable
-                self.global_endpoint_manager.mark_endpoint_unavailable_for_read(self.request.location_endpoint_to_route, True)
+                self.global_endpoint_manager.mark_endpoint_unavailable_for_read(
+                    self.request.location_endpoint_to_route,
+                    True)
             else:
-                self.global_endpoint_manager.mark_endpoint_unavailable_for_write(self.request.location_endpoint_to_route, True)
+                self.global_endpoint_manager.mark_endpoint_unavailable_for_write(
+                    self.request.location_endpoint_to_route,
+                    True)
 
         # set the refresh_needed flag to ensure that endpoint list is
         # refreshed with new writable and readable locations