+++ sdk/cosmos/azure-cosmos/azure/cosmos/aio/_retry_utility_async.py
@@ -45,7 +45,7 @@ from ..http_constants import HttpHeaders, StatusCodes, SubStatusCodes
 
 # pylint: disable=protected-access, disable=too-many-lines, disable=too-many-statements, disable=too-many-branches
 
-async def ExecuteAsync(client, global_endpoint_manager, function, *args, **kwargs):
+async def ExecuteAsync(client, global_endpoint_manager, function, *args, **kwargs): # pylint: disable=too many local variables
     """Executes the function with passed parameters applying all retry policies
 
     :param object client: